Finish error handling for now
The "for now" is that the way it's interfaced with at the top level is
kind of grody (lex_input() returns a boolean but execute_bytecode
throws?) and there's no way for the user to control it.

I'm not terribly worried about the repl grodiness because the way that
code is invoked is due for a rethink later on anyways.