Rebuild all nf modules after the first dirty one
There's no dependency graph for builtins/, so we can't tell what modules
a given module depends on. So, if we have to rebuild a module, we play
it safe and assume that every module after that one also needs to be
rebuilt, to pick up (e.g.) new constant definitions.